Improve pickling API and perf.
- Introduce UnpickledSize which returns the expected size of a pickled
item once unpickled.

- Introduce an Unpickle overload which takes a Span<byte> and one that
takes an IBufferWriter for output. Both of these enable the caller to
reuse buffers across multiple unpicklings, which increases perf and
reduces pressure on the GC.

- Introduce a Pickle overload which takes an IBufferWriter as output.
Using this overload enables the caller to reuse buffers, and eliminates
a full data copy by unpickling directly into the final destination
buffer.

- The code was rewritten to mostly be safe by leveraging Span<T>. There
is only a teeny bit of unsafe code left in the pickling path.